fix: ui row color change & auction advice
